Refactor img_generator.py into separate modules
This commit refactors the main `img_generator.py` script by splitting it into three distinct files to improve modularity and maintainability:

1.  `config.py`: Contains all user-configurable parameters, such as image dimensions, model names, and API settings.
2.  `prompts.py`: Holds the logic for generating random prompts, including all the lists of terms and themes.
3.  `img_generator.py`: The main script, which now contains only the core logic for process management and image generation, importing the configuration and prompts from the other modules.
